Added verifySelection page
Added the html page that will verify a user’s restaurant selection.

The page has text fields to display the name of the establishment, a
picture from Google Maps of the location, address, distance from
current location, and when the voucher will expire.

There are confirm and cancel buttons at the bottom of the page.