Updated the Create Account Page
Updated signup.html so that when its “create user” account is clicked,
it will 1) make sure the data being submitted is valid 2) Create a new
user 3) Add the additional information to that user.

There is a bug in the commented out code in the submit function.
Firebase is creating a new user, which we want to pass into
writeUserData() to add more information to the user. However,
writeUserData() is being called before Firebase finishes creating the
user, resulting in the program sending in a null value as the user,
which makes Firebase throw an error.

Pushing this with the error so that another team member can look at its
asynchronization issues.